pyproject.toml
[tool.poetry]
name = "crud"
version = "0.1.0"
description = ""
authors = ["alireza <alirezaghanati@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
requests = "^2.31.0"
django-dotenv = "^1.4.2"
psycopg2-binary = "^2.9.9"
django-storages = "^1.14.2"
boto3 = "^1.34.54"
django-filter = "^23.5"
django-split-settings = "^1.2.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
platformdirs = "^4.2.0"
colorlog = "^6.8.2"
pandas = "^2.2.1"
openpyxl = "^3.1.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
